+++ original/Code/Commando/dlgevaobjectivestab.h
@@ -37,7 +37,7 @@ private:
 	////////////////////////////////////////////////////////////////
 	//	Static methods
 	////////////////////////////////////////////////////////////////
-	static int CALLBACK	ListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uint32_t user_param);
+	static int CALLBACK	ListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uintptr_t user_param);
 
 	////////////////////////////////////////////////////////////////
 	//	Private member data
