+++ original/Code/Commando/dlgsavegame.h
@@ -49,7 +49,7 @@ private:
 	////////////////////////////////////////////////////////////////
 	//	Static members
 	////////////////////////////////////////////////////////////////
-	static int CALLBACK LoadListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uint32_t user_param);
+	static int CALLBACK LoadListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uintptr_t user_param);
 
 	////////////////////////////////////////////////////////////////
 	//	Private methods
