+++ original/Code/wwui/listctrl.h
@@ -21,7 +21,7 @@ class TextureClass;
 ////////////////////////////////////////////////////////////////
 //	Typedefs
 ////////////////////////////////////////////////////////////////
-typedef int (CALLBACK *LISTCTRL_SORT_CALLBACK) (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uint32_t user_param);
+typedef int (CALLBACK *LISTCTRL_SORT_CALLBACK) (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uintptr_t user_param);
 
 ////////////////////////////////////////////////////////////////
 //
@@ -81,9 +81,9 @@ public:
 	bool				Set_Entry_Text (int index, int col_index, const WCHAR *text);
 	bool				Set_Entry_Int (int index, int col_index, int value);
 	bool				Set_Entry_Color (int index, int col_index, const Vector3 &color);
-	bool				Set_Entry_Data (int index, int col_index, uint32_t user_data);
+	bool				Set_Entry_Data (int index, int col_index, uintptr_t user_data);
 	bool				Select_Entry (int index, bool onoff);
-	uint32_t			Get_Entry_Data (int index, int col_index);
+	uintptr_t			Get_Entry_Data (int index, int col_index);
 	const WCHAR *	Get_Entry_Text (int index, int col_index);
 	bool				Is_Entry_Selected (int index);
 	bool				Delete_Entry (int index);
@@ -120,7 +120,7 @@ public:
 	//
 	//	Sort support
 	//
-	void				Sort (LISTCTRL_SORT_CALLBACK sort_callback, uint32_t user_param);
+	void				Sort (LISTCTRL_SORT_CALLBACK sort_callback, uintptr_t user_param);
 	void				Set_Sort_Designator (int col_index, SORT_TYPE type);
 	void				Sort_Alphabetically (int col_index, SORT_TYPE type);
 
@@ -163,7 +163,7 @@ protected:
 	void				Update_Scroll_Pos (void);
 	void				Update_Scroll_Bar_Visibility (void);
 	void				Update_Row_Height (int row_index);
-	void				Quick_Sort (int start_index, int end_index, LISTCTRL_SORT_CALLBACK sort_callback, uint32_t user_param);
+	void				Quick_Sort (int start_index, int end_index, LISTCTRL_SORT_CALLBACK sort_callback, uintptr_t user_param);
 	int				Find_Last_Page_Top_Entry (void);
 	void				Scroll_Page (int direction);
 	int				Find_End_Of_Page (void);
@@ -171,7 +171,7 @@ protected:
 	
 	void				Render_Entry (const RectClass &rect, int col_index, int row_index);
 
-	static int CALLBACK	Default_Sort_Callback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uint32_t user_param);
+	static int CALLBACK	Default_Sort_Callback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uintptr_t user_param);
 
 	////////////////////////////////////////////////////////////////
 	//	Protected constants
@@ -265,8 +265,8 @@ public:
 	//
 	//	User data access
 	//
-	uint32_t				Get_User_Data (void) const			{ return UserData; }
-	void					Set_User_Data (uint32_t user_data)	{ UserData = user_data; }
+	uintptr_t				Get_User_Data (void) const			{ return UserData; }
+	void					Set_User_Data (uintptr_t user_data)	{ UserData = user_data; }
 
 	//
 	//	Icon support
@@ -283,7 +283,7 @@ private:
 	////////////////////////////////////////////////////////////////
 	WideStringClass						Name;
 	Vector3									Color;
-	uint32_t									UserData;
+	uintptr_t									UserData;
 	DynamicVectorClass<StringClass>	IconList;
 };
 
@@ -348,8 +348,8 @@ public:
 	void					Set_Entry_Color (int index, const Vector3 &color)	{ EntryList[index]->Set_Color (color); }
 	const Vector3 &	Get_Entry_Color (int index) const						{ return EntryList[index]->Get_Color (); }
 
-	void					Set_Entry_Data (int index, uint32_t data)				{ EntryList[index]->Set_User_Data (data); }
-	uint32_t				Get_Entry_Data (int index) const							{ return EntryList[index]->Get_User_Data (); }
+	void					Set_Entry_Data (int index, uintptr_t data)				{ EntryList[index]->Set_User_Data (data); }
+	uintptr_t				Get_Entry_Data (int index) const							{ return EntryList[index]->Get_User_Data (); }
 
 	int					Get_Icon_Count (int index) const							{ return EntryList[index]->Get_Icon_Count (); }
 	const char *		Get_Icon (int index, int icon_index)					{ return EntryList[index]->Get_Icon (icon_index); }
